refactor(telemetry): point DefaultPricing at domain config (#1350)
Re-point the telemetry package's DefaultPricing dependency from the
infrastructure re-export shim to the domain function directly, and
remove the now-orphaned infra shim and its dead white-box test.

Refs #1350 (item 1)
